Folder renamed and component done
